+++ test/2.0/test-middleware-swagger-validator.js
@@ -1251,5 +1251,62 @@ describe('Swagger Validator Middleware v2.0', function () {
           .end(helpers.expectContent(expectedMessage, done));
       });
     });
+
+    describe('should handle allowEmptyValue (Issue 282)', function () {
+      it('allowEmptyValue false', function (done) {
+        var cPetStore = _.cloneDeep(petStoreJson);
+        var expectedMessage = 'Request validation failed: Parameter (arg0) is not a valid integer: ';
+
+        cPetStore.paths['/pets/{id}'].get.parameters = [{
+            in: 'query',
+          name: 'arg0',
+          type: 'integer'
+        }];
+
+        helpers.createServer([cPetStore], {}, function (app) {
+          request(app)
+            .get('/api/pets/1')
+            .query({
+              arg0: ''
+            })
+            .expect(400)
+            .end(helpers.expectContent(expectedMessage, done));
+        });
+      });
+
+      it('allowEmptyValue true', function (done) {
+        var cPetStore = _.cloneDeep(petStoreJson);
+
+        cPetStore.paths['/pets/{id}'].get.parameters = [{
+            in: 'query',
+          name: 'arg0',
+          type: 'integer',
+          allowEmptyValue: true
+        }];
+
+        helpers.createServer([cPetStore], {
+          swaggerRouterOptions: {
+            controllers: {
+              getPetById: function (req, res) {
+                res.end('OK');
+              }
+            }
+          }
+        }, function (app) {
+          try {
+            request(app)
+              .get('/api/pets/1')
+              .query({
+                arg0: ''
+              })
+              .expect(200)
+              .end(helpers.expectContent('OK', done));
+          } catch (err) {
+            console.log(err.stack);
+            done();
+          }
+        });
+      });
+    });
   });
 });